pyproject.toml
[project]
name = "tinymmo-loginserver"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic (>=2.11.4,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "redis (>=6.0.0,<7.0.0)"
]

[tool.poetry]
packages = [{include = "login_server"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
